Committ archivos Wilson mas algunas correciones

+++ php/views/series/series.php
@@ -1,10 +1,30 @@
 <!-- Header -->
 <?php include "header.php"?>
  
+<script>
+    function confirmacion() {
+        var respuesta = confirm("¿Desea realmente borrar el registro?");
+        if (respuesta == true) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+</script>
+
   <div class="container">
     <h1 class="text-center" >Operaciones CRUD Series</h1>
       <a href="create.php" class='btn btn-outline-dark mb-2'> <i class="bi bi-person-plus"></i> Crear nueva serie</a>
- 
+
+      <?php
+           require_once('../../controllers/serie_controller.php');
+           require_once('../../controllers/director_controller.php');
+           require_once('../../controllers/platform_controller.php');
+           
+           $series = getSeries();
+
+           if(count($series)> 0) {
+?>      
         <table class="table table-striped table-bordered table-hover">
           <thead class="table-dark">
             <tr>
@@ -20,12 +40,8 @@
           </thead>
             <tbody>
               <tr>
+        <?php
   
-          <?php
-           require_once('../../controllers/serie_controller.php');
-           require_once('../../controllers/director_controller.php');
-           require_once('../../controllers/platform_controller.php');
-           $series = getSeries();
            foreach($series as $serie){
             $directorName = '';
             $platformName = '';
@@ -52,17 +68,29 @@
             echo " <td >{$actores}</td>";
             //echo " <td class='text-center'> <a href='view.php?user_id={$serie->getId()}' class='btn btn-primary'> <i class='bi bi-eye'></i> View</a> </td>";
             echo " <td class='text-center' > <a href='update.php?edit&serie_id={$serie->getId()}' class='btn btn-primary'><i class='bi bi-pencil'></i> EDITAR</a> </td>";
-            echo " <td  class='text-center'>  <a href='delete.php?delete={$serie->getId()}' class='btn btn-danger'> <i class='bi bi-trash'></i> BORRAR</a> </td>";
+            echo " <td  class='text-center'>  <a onclick='return confirmacion()' href='delete.php?delete={$serie->getId()}'  class='btn btn-danger'> <i class='bi bi-trash'></i> BORRAR</a> </td>";
             echo " </tr> ";
            }
                 ?>
               </tr>  
             </tbody>
         </table>
+        <?php
+        } else {
+    ?>
+
+    <div class="alert alert-warning" role="alert">
+      Aún no existen series.
+    </div> 
+
+    <?php
+        }
+    ?>
+
   </div>
  
 <div class="container text-center mt-5">
-      <a href="../../index.html" class="btn btn-warning mt-5"> Regresar </a>
+      <a href="../../../index.html" class="btn btn-warning mt-5"> Regresar </a>
     <div>
  
 <!-- Footer -->
